start moving to convolutions.jl
move nextfastfft to convolutions.jl
