+++ ext/OffsetArraysExt.jl
@@ -2,6 +2,6 @@ module OffsetArraysExt
 import DSP
 import OffsetArrays
 
-DSP.conv_axis_with_offset(::OffsetArrays.IdOffsetRange) = true
+DSP.Convolutions.conv_axis_with_offset(::OffsetArrays.IdOffsetRange) = true
 
 end
