+++ src/DSP.jl
@@ -1,19 +1,19 @@
 module DSP
 
-using FFTW
-using LinearAlgebra: Transpose, mul!, rmul!
-using IterTools: subsets
+using LinearAlgebra: mul!
 
-export conv, conv!, deconv, filt, filt!, xcorr
+export deconv, filt, filt!, xcorr
 
 # This function has methods added in `periodograms` but is not exported,
 # so we define it here so one can do `DSP.allocate_output` instead of
 # `DSP.Periodograms.allocate_output`.
 function allocate_output end
 
-include("dspbase.jl")
+include("convolutions.jl")
 
+include("dspbase.jl")
 include("util.jl")
+
 include("unwrap.jl")
 include("windows.jl")
 include("periodograms.jl")
@@ -23,7 +23,7 @@ include("estimation.jl")
 include("diric.jl")
 
 using Reexport
-@reexport using .Util, .Windows, .Periodograms, .Filters, .LPC, .Unwrap, .Estimation
+@reexport using .Util, .Windows, .Periodograms, .Filters, .LPC, .Unwrap, .Estimation, .Convolutions
 
 include("deprecated.jl")
 end
