+++ src/util.jl
@@ -4,6 +4,7 @@ import Base: *
 using LinearAlgebra: mul!, BLAS
 using FFTW
 using Statistics: mean
+using ..Convolutions: nextfastfft
 
 export  hilbert,
         fftintype,
@@ -102,37 +103,6 @@ fftouttype(::Type{T}) where {T<:Union{Real,Complex}} = ComplexF64
 fftabs2type(::Type{Complex{T}}) where {T<:FFTW.fftwReal} = T
 fftabs2type(::Type{T}) where {T<:FFTW.fftwReal} = T
 fftabs2type(::Type{T}) where {T<:Union{Real,Complex}} = Float64
-
-# Get next fast FFT size for a given signal length
-const FAST_FFT_SIZES = (2, 3, 5, 7)
-
-"""
-    nextfastfft(n::Integer)
-    nextfastfft(ns::Tuple)
-
-Return an estimate for the padded size of an array that
-is optimal for the performance of an n-dimensional FFT.
-
-For `Tuple` inputs, this returns a `Tuple` that is the size
-of the padded array.
-
-For `Integer` inputs, this returns the closest product of 2, 3, 5, and 7
-greater than or equal to `n`.
-FFTW contains optimized kernels for these sizes and computes
-Fourier transforms of inputs that are a product of these
-sizes faster than for inputs of other sizes.
-
-!!! warning
-
-    Currently, `nextfastfft(ns::Tuple)` is implemented as
-    `nextfastfft.(ns)`. This may change in future releases if
-    a better estimation model is found.
-
-    It is recommended to use `nextfastfft.(ns)` to obtain
-    a tuple of padded lengths for 1D FFTs.
-"""
-nextfastfft(n::Integer) = nextprod(FAST_FFT_SIZES, n)
-nextfastfft(ns::Tuple{Vararg{Integer}}) = nextfastfft.(ns)
 
 
 ## COMMON DSP TOOLS
