+++ test/dsp.jl
@@ -4,7 +4,7 @@ using Test, OffsetArrays
 using DSP: filt, filt!, deconv, conv, conv!, xcorr,
            optimalfftfiltlength, unsafe_conv_kern_os!, _conv_kern_fft!,
            nextfastfft
-
+using DSP.Convolutions: optimalfftfiltlength, unsafe_conv_kern_os!, _conv_kern_fft!
 
 
 @testset "filt" begin
